Change manual unroll of enum enumeration to loop
This will clear issues that would be caused by improper order of
tool enumerations, by instead making it so that all enums are created
and added in the same order that they are listed no matter what.
This also removes redundant code within EToolType by giving the file
directories meaning.